+++ packages/atomic/src/components/search/facets/atomic-color-facet/atomic-color-facet.tsx
@@ -213,7 +213,7 @@ export class AtomicColorFacet
     this.facet = buildFacet(this.bindings.engine, {options});
     this.facetId = this.facet.state.facetId;
     const facetInfo: FacetInfo = {
-      label: this.label,
+      label: () => this.bindings.i18n.t(this.label),
       facetId: this.facetId!,
       element: this.host,
     };