+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.116.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.115.0...@coveo/atomic@1.116.0) (2022-10-07)
+
+
+### Features
+
+* **atomic:** add calendar time support + result-timespan component ([#2475](https://github.com/coveo/ui-kit/issues/2475)) ([59ff2a8](https://github.com/coveo/ui-kit/commit/59ff2a8ba0b52fd1afd9af0ba50521c42cdd1c76))
+
+
+
+
+
+# [1.115.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.114.0...@coveo/atomic@1.115.0) (2022-10-06)
+
+
+### Features
+
+* **atomic,headless:** add rec result & template components, add missing rec state, add recommendations list styling & layout ([#2484](https://github.com/coveo/ui-kit/issues/2484)) ([c3fb9eb](https://github.com/coveo/ui-kit/commit/c3fb9ebab3181da7b55a3c8152822455399fbfa1))
+* **atomic:** add atomic-result-localized-text component ([#2483](https://github.com/coveo/ui-kit/issues/2483)) ([662dc14](https://github.com/coveo/ui-kit/commit/662dc14e1164e62ed2ae3c8356c610f90d0b8969))
+
+
+
+
+
 # [1.114.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.113.9...@coveo/atomic@1.114.0) (2022-10-05)